[AMQCPP-150] When performing tight unmarshalling of compressed long long values (e.g. int and short), cast the values to unsigned, so that the sign bit will not appear in the resulting long
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@589089 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/activemq/connector/openwire/marshal/BaseDataStreamMarshaller.cpp
@@ -409,13 +409,13 @@ long long BaseDataStreamMarshaller::tightUnmarshalLong(
             if( bs->readBoolean() ) {
                 return dataIn->readLong();
             } else {
-                return dataIn->readInt();
+                return (unsigned int)dataIn->readInt();
             }
 
         } else {
 
             if( bs->readBoolean()) {
-                return dataIn->readShort();
+                return (unsigned short)dataIn->readShort();
             } else {
                 return 0;
             }
